feat(api): move #authenticateApplication usecase to identity access management context
